% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dative_csv.R
\name{read_dative_csv}
\alias{read_dative_csv}
\title{Reads CSV export from dative}
\usage{
read_dative_csv(file_path)
}
\arguments{
\item{file_path}{the path to the CSV file}
}
\value{
a tibble with the relevant columns from the CSV file (and ID column first)
}
\description{
Reads CSV export from dative
}
