% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_brackets.R
\name{strip_brackets}
\alias{strip_brackets}
\title{Strip Brackets From the Ends of a String}
\usage{
strip_brackets(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector with any straight brackets or forward slashes removed from the beginning and end of the string.
}
\description{
Strip Brackets From the Ends of a String
}
\examples{
x <- "/mob/"
strip_brackets(x)
}
